+++ App/ATFGenerator/src/AutomaticTransferFunction/ATFGenerator.h
@@ -146,7 +146,7 @@ public:
   /// </summary>
   void GenerateLaplacianSummedHistogram();
 
-  void GenerateDataValuesFile(float *x, unsigned char *v, int n);
+  void GenerateDataValuesFile(int *x, unsigned char *v, int n);
 
   /// <summary>
   /// Gets the transfer function.
@@ -211,7 +211,7 @@ private:
   /// </summary>
   /// <returns>Returns a float array with the distances associated 
   /// to all 256 values, ordered by value.</returns>
-  float GetBoundaryDistancies(float* x, unsigned char *v, int *n);
+  float GetBoundaryDistancies(int* x, unsigned char *v, int *n);
 
   unsigned int GetId(unsigned int x, unsigned int y, unsigned int z);
